requirements.txt
torch
detectron2
easyocr
opencv-python
numpy
Pillow
openai